+++ cmd/provisioner-localpv/app/helper.go
@@ -43,6 +43,26 @@ var (
 	CmdTimeoutCounts = 120
 )
 
+// HelperPodOptions contains the options that
+// will be extracted from the persistent volume
+type HelperPodOptions struct {
+	cmdsForPath []string
+	name        string
+	path        string
+	nodeName    string
+}
+
+// validate checks that the required fields to launch
+// helper pods are valid. helper pods are used to either
+// create or delete a directory (path) on a given node (nodeName).
+// name refers to the volume being created or deleted.
+func (pOpts *HelperPodOptions) validate() error {
+	if pOpts.name == "" || pOpts.path == "" || pOpts.nodeName == "" {
+		return errors.Errorf("invalid empty name or path or node")
+	}
+	return nil
+}
+
 // getPathAndNodeForPV inspects the PV spec to determine the hostpath
 //  and the node of OpenEBS Local PV. Both types of OpenEBS Local PV
 //  (storage type = hostpath and device) use:
